Fix id type

+++ datasource/src/main/java/com/bruno13palhano/UserData.java
@@ -3,10 +3,10 @@ package com.bruno13palhano;
 public interface UserData<T extends User> {
     void insert(T user);
     void update(T user);
-    void delete(Long userId);
+    void delete(Integer userId);
     T getByUsername(String username);
     Boolean usernameAlreadyExist(String username);
     Boolean emailAlreadyExist(String email);
     T getByEmail(String email);
-    T getById(Long userId);
+    T getById(Integer userId);
 }
